mDNSResponder-2200.100.94.0.2
Imported from mDNSResponder-2200.100.94.0.2.tar.gz

+++ ServiceRegistration/thread-service.c
@@ -149,7 +149,7 @@ thread_service_note(const char *owner_id, thread_service_t *tservice, const char
             struct thread_unicast_service *service = &tservice->u.unicast;
             uint16_t port;
 
-            port = (service->port[0] << 8) | service->port[1];
+            port = (uint16_t)(service->port[0] << 8) | service->port[1];
             SEGMENTED_IPv6_ADDR_GEN_SRP(&service->address, service_add_buf);
             INFO(PUB_S_SRP " SRP service " PRI_SEGMENTED_IPv6_ADDR_SRP "%%%d, rloc16 %x " PUB_S_SRP, owner_id,
                  SEGMENTED_IPv6_ADDR_PARAM_SRP(&service->address, service_add_buf),
